cleanup: merge `AccessIndicator` & `AccessIndicatorUI` files
