--- src/components/[guild]/RoleCard/components/AccessIndicator/index.ts
@@ -1,3 +1,0 @@
-import { AccessIndicator } from "./AccessIndicator"
-
-export { AccessIndicator }
